requirements.txt
fastmcp>=0.9.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic>=2.5.0
python-dotenv>=1.0.0
structlog>=23.2.0
tenacity>=8.2.0
